Refactor: 각 form 마다 customhook 구현
